Refactor monolithic test_filler into unit tests.
